Fix issue with golang dependencies due to semver format in module

+++ nasConvert/NetWorkName.go
@@ -20,7 +20,7 @@ func FullNetworkNameToNas(name string) (fullNetworkName nasType.FullNameForNetwo
 		if i == 0 {
 			buf = append(buf, char)
 		} else {
-			buf[i-1] = (buf[i-1] & nasType.GetBitMask(idx+1, 0)) + uint8(char<<idx)
+			buf[i-1] = (buf[i-1] & nasType.GetBitMask(idx+1, 0)) + char<<idx
 			buf = append(buf, char>>(8-idx))
 			idx--
 			// if idx overflow, it will round to max(uint8) == 255 == ^uint8(0)
@@ -49,7 +49,7 @@ func ShortNetworkNameToNas(name string) (shortNetworkName nasType.ShortNameForNe
 		if i == 0 {
 			buf = append(buf, char)
 		} else {
-			buf[i-1] = (buf[i-1] & nasType.GetBitMask(idx+1, 0)) + uint8(char<<idx)
+			buf[i-1] = (buf[i-1] & nasType.GetBitMask(idx+1, 0)) + char<<idx
 			buf = append(buf, char>>(8-idx))
 			idx--
 			// if idx overflow, it will round to max(uint8) == 255 == ^uint8(0)
